refactor(wallet): split hardware wallet tests
