+++ packages/wallet/package.json
@@ -49,7 +49,8 @@
     "prepack": "yarn build",
     "test": "jest -c ./jest.config.js",
     "test:build:verify": "tsc --build ./test",
-    "test:hw": "jest ./test/hardware/* -c ./hw.jest.config.js --runInBand",
+    "test:hw:ledger": "jest ./test/hardware/ledger/* -c ./hw.jest.config.js --runInBand",
+    "test:hw:trezor": "jest ./test/hardware/trezor/* -c ./hw.jest.config.js --runInBand",
     "test:debug": "DEBUG=true yarn test",
     "test:e2e": "shx echo 'test:e2e' command not implemented yet"
   },
