+++ packages/wallet/test/hardware/ledger/LedgerKeyAgent.integration.test.ts
@@ -2,7 +2,7 @@ import * as Crypto from '@cardano-sdk/crypto';
 import { CML, Cardano } from '@cardano-sdk/core';
 import { CommunicationType, KeyAgent, util } from '@cardano-sdk/key-management';
 import { LedgerKeyAgent } from '@cardano-sdk/hardware-ledger';
-import { ObservableWallet, SingleAddressWallet, restoreKeyAgent, setupWallet } from '../../src';
+import { ObservableWallet, SingleAddressWallet, restoreKeyAgent, setupWallet } from '../../../src';
 import { createStubStakePoolProvider } from '@cardano-sdk/util-dev';
 import { firstValueFrom } from 'rxjs';
 import { dummyLogger as logger } from 'ts-log';
@@ -13,7 +13,7 @@ import {
   mockRewardsProvider,
   mockTxSubmitProvider,
   mockUtxoProvider
-} from '../mocks';
+} from '../../mocks';
 
 const createWallet = async (keyAgent: KeyAgent) => {
   const txSubmitProvider = mockTxSubmitProvider();
